Refactor, add drawAsset to FloorMapRenderer
