+++ features/il_display_assets_map/lib/src/widgets/assets_painter.dart
@@ -23,56 +23,14 @@ class AssetsPainter extends CustomPainter {
   void paint(Canvas canvas, Size size) {
     floorMapRenderer.drawZoneLabels(canvas, floorMap);
 
-    for (var asset in assets) {
-      drawAsset(asset, canvas);
-    }
-  }
-
-  void drawAsset(Asset asset, Canvas canvas) {
-    const double assetCircleRadius = 4.0;
-    const double labelBottomMargin = 10;
-    const double labelFontSize = 16;
-
-    // transform asset position
-
-    double scale = transform.getMaxScaleOnAxis();
-    var translation = transform.getTranslation();
-
     Rect trackingArea = floorMap.trackingArea;
 
-    double ax = asset.x + trackingArea.left;
-    double ay = asset.y + trackingArea.top;
+    for (var asset in assets) {
+      double ax = asset.x + trackingArea.left;
+      double ay = asset.y + trackingArea.top;
 
-    double tx = ax * scale + translation.x;
-    double ty = ay * scale + translation.y;
-    Offset assetPosition = Offset(tx, ty);
-
-    // draw position circle
-
-    Paint paint = Paint();
-    paint.color = AppColors.primaryBlueColor;
-    paint.style = PaintingStyle.fill;
-
-    canvas.drawCircle(assetPosition, assetCircleRadius, paint);
-
-    // draw asset name
-
-    var textPainter = TextPainter(
-      textDirection: TextDirection.ltr,
-      text: TextSpan(
-        text: asset.name,
-        style: const TextStyle(color: Colors.white, fontSize: labelFontSize),
-      ),
-    );
-
-    textPainter.layout();
-
-    var labelPosition = assetPosition - Offset(textPainter.width / 2, textPainter.height + labelBottomMargin);
-    var labelRectPosition = labelPosition - const Offset(12, 2);
-    var labelRectSize = Offset(textPainter.width + 24, textPainter.height + 4);
-
-    floorMapRenderer.drawRRect(canvas, labelRectPosition, labelRectSize, labelRectSize.dy / 2, paint);
-    textPainter.paint(canvas, labelPosition);
+      floorMapRenderer.drawAsset(canvas, asset.name, Offset(ax, ay));
+    }
   }
 
   @override
